Update v.0.9.0 (#22)
* Removed some console logs (maybe its the case)

* More fixes
Removed the janky setInterval, added events to checkboxes and whatnot. Added a toggle for custom theme. Custom theme imports work perfectly and so does saving for evrey setting; Only thing that doesn't work is if you try to import the same custom theme file twice but that's just how it works because we're listening for a change on CustomThemeImport

* Update v0.9.0 (I didnt feel like making another branch)

* Update v.0.9.0 (No permission edition)

* readme keybinds

* readame keybinds

* Custom snap is now not the same as snap denominator